requirements.txt
tensorflow == 1.13.1
opencv-python == 4.1.0.25
numpy == 1.16.4
jupyter == 1.0.0
Keras == 2.2.4
plaidml-keras == 0.6.0
Pillow == 6.0.0
scikit-image == 0.15.0
scipy == 1.3.0
PyAutoGUI == 0.9.42
